参数属性类型修改，工具类方法修改

+++ jeecg-boot/jeecg-boot-module-system/src/main/java/org/jeecg/modules/game/entity/DailyGiftPackageBuyVO.java
@@ -35,7 +35,7 @@ public class DailyGiftPackageBuyVO implements Serializable {
 	/**
 	 * 玩家id
 	 */
-    private Integer playerId;
+    private Long playerId;
 	
 	/**
 	 * 礼包id
